requirements_pip.txt
asn1crypto==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
click==7.1.1
community==1.0.0b1
cryptography==2.5
cycler==0.10.0
decorator==4.4.0
dill==0.2.9
Flask==1.1.1
html5lib==1.0.1
idna==2.8
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.0.1
jupyter-core==4.6.3
keepalive==0.5
kiwisolver==1.0.1
klepto==0.1.6
MarkupSafe==1.1.1
matplotlib==3.0.3
nbformat==4.4.0
networkx==2.2
numpy==1.16.2
olefile==0.46
pandas==0.24.2
Pillow==6.0.0
plyfile==0.7
pox==0.2.5
protobuf==3.7.1
pycparser==2.20
pygraphviz==1.5
#PyMetis==2019.1.1
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.15.2
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2018.9
rdflib==4.2.2
requests==2.21.0
retrying==1.3.3
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
SPARQLWrapper==1.8.2
tensorboardX==1.6
#torch==1.4.0
torch-cluster==1.2.4
torch-geometric==1.1.2
torch-scatter==1.1.2
torch-sparse==0.2.4
torch-spline-conv==1.0.6
tornado==6.0.4
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==1.0.1
